+++ app/src/main/java/com/lumiyaviewer/lumiya/slproto/types/LLQuaternion.kt
@@ -90,7 +90,7 @@ class LLQuaternion {
         this.inverseMatrix = null
         Float f = fArr[0] + 1.0f + fArr[5] + fArr[10]
         if (f > 0.5f) {
-            Float sqrt = (Float) (Math.sqrt((Double) f) * 2.0d)
+            Float sqrt = (Float) (Math.sqrt(f.toDouble()) * 2.0d)
             this.x = (fArr[9] - fArr[6]) / sqrt
             this.y = (fArr[2] - fArr[8]) / sqrt
             this.z = (fArr[4] - fArr[1]) / sqrt
@@ -177,7 +177,7 @@ class LLQuaternion {
         Float f4 = byteBuffer.getFloat()
         Float f5 = 1.0f - (((f2 * f2) + (f3 * f3)) + (f4 * f4))
         if (f5 > 0.0f) {
-            f = (Float) Math.sqrt((Double) f5)
+            f = Math.sqrt(f5.toDouble()).toFloat()
         }
         return LLQuaternion(f2, f3, f4, f)
     }
@@ -214,7 +214,7 @@ class LLQuaternion {
             return LLQuaternion(lLVector36.x, lLVector36.y, lLVector36.z, 0.0f)
         }
         LLQuaternion lLQuaternion = LLQuaternion()
-        lLQuaternion.setQuat((Float) Math.acos((Double) dot), cross)
+        lLQuaternion.setQuat(Math.toFloat().acos(dot.toDouble()), cross)
         return lLQuaternion
     }
 
@@ -225,7 +225,7 @@ class LLQuaternion {
         Float f3 = lLVector3.y
         Float f4 = lLVector3.z
         if (magVecSquared > 0.0f) {
-            f = (Float) Math.sqrt((Double) magVecSquared)
+            f = Math.sqrt(magVecSquared.toDouble()).toFloat()
         }
         return LLQuaternion(f2, f3, f4, f)
     }
@@ -250,9 +250,9 @@ class LLQuaternion {
         if (f2 >= -1.0f) {
             f = f2
         }
-        Float sqrt = (Float) Math.sqrt((Double) (1.0f - (f * f)))
+        Float sqrt = Math.sqrt((1.0f - (f * f.toDouble()).toFloat()))
         Float f3 = Math.abs(sqrt) < 5.0E-4f ? 1.0f : 1.0f / sqrt
-        Float acos = ((Float) Math.acos((Double) f)) * 2.0f
+        Float acos = (Math.toFloat().acos(f.toDouble())) * 2.0f
         if (acos > 3.1415927f) {
             lLVector3.x = (-this.x) * f3
             lLVector3.y = (-this.y) * f3
@@ -328,7 +328,7 @@ class LLQuaternion {
     }
 
     Float normalize() {
-        Float sqrt = (Float) Math.sqrt((Double) ((this.x * this.x) + (this.y * this.y) + (this.z * this.z) + (this.w * this.w)))
+        Float sqrt = Math.sqrt(((this.x * this.x.toDouble()).toFloat() + (this.y * this.y) + (this.z * this.z) + (this.w * this.w)))
         if (sqrt <= 1.0E-7f) {
             this.x = 0.0f
             this.y = 0.0f
@@ -384,11 +384,11 @@ class LLQuaternion {
         LLVector3 lLVector3 = LLVector3(f2, f3, f4)
         lLVector3.normVec()
         Float f5 = 0.5f * f
-        Float sin = (Float) Math.sin((Double) f5)
+        Float sin = Math.toFloat().sin(f5.toDouble())
         this.x = lLVector3.x * sin
         this.y = lLVector3.y * sin
         this.z = lLVector3.z * sin
-        this.w = (Float) Math.cos((Double) f5)
+        this.w = Math.toFloat().cos(f5.toDouble())
         normalize()
         this.matrix = null
         this.inverseMatrix = null
@@ -398,11 +398,11 @@ class LLQuaternion {
         LLVector3 lLVector32 = LLVector3(lLVector3)
         lLVector32.normVec()
         Float f2 = 0.5f * f
-        Float sin = (Float) Math.sin((Double) f2)
+        Float sin = Math.toFloat().sin(f2.toDouble())
         this.x = lLVector32.x * sin
         this.y = lLVector32.y * sin
         this.z = lLVector32.z * sin
-        this.w = (Float) Math.cos((Double) f2)
+        this.w = Math.toFloat().cos(f2.toDouble())
         normalize()
         this.matrix = null
         this.inverseMatrix = null
