+++ app/src/main/java/com/lumiyaviewer/lumiya/slproto/types/LLVector2.kt
@@ -53,7 +53,7 @@ class LLVector2 {
     }
 
     Float magVec() {
-        return (Float) Math.sqrt((Double) ((this.x * this.x) + (this.y * this.y)))
+        return Math.sqrt(((this.x * this.x.toDouble()).toFloat() + (this.y * this.y)))
     }
 
     Unit mul(Float f) {
@@ -62,7 +62,7 @@ class LLVector2 {
     }
 
     Float normVec() {
-        Float sqrt = (Float) Math.sqrt((Double) ((this.x * this.x) + (this.y * this.y)))
+        Float sqrt = Math.sqrt(((this.x * this.x.toDouble()).toFloat() + (this.y * this.y)))
         if (sqrt > 1.0E-7f) {
             Float f = 1.0f / sqrt
             this.x *= f
